[qt] Generalise ImportCurrencyDialog into ImportEntityDialog
Qt's moc can't process class templates with Q_OBJECT, so this is a
single concrete QDialog driven by type-erased preview rows and
callbacks (label_of/import_one) rather than a template parametrized on
the entity type. Currency's importFromXML() now builds its preview
rows and import callback and passes them in, instead of owning a
currency-specific dialog. ImportCurrencyDialog is deleted.

Steps 1-2 of the CSV/XML qt-profile capability plan; step 3 (the
has_csv_xml_io mustache/codegen wiring) follows.

Story-ID: 7F9C0F29-4268-4B62-A1BB-2153ECF0A858
Task-ID: 2EAF2F99-B71F-410D-985A-7A616E40CB88
