+++ doc/agile/versions/v0/sprint_22/commission_currency/story.org
@@ -102,7 +102,7 @@ auxiliary type (=rounding_type=, =monetary_nature=, =currency_market_tier=).
 | [[id:3CE7779C-7D20-4203-A4E9-1FBDBFAEF537][File Wt and HTTP gap captures for currency]] | BACKLOG | | | Identify Wt UI and HTTP handler gaps for currency discovered during commissioning and file them as backlog captures for future work. |
 | [[id:F8A8163A-BC1F-4702-B920-1002ED764CEF][Migrate currency model from ores.codegen.table to ores.codegen.entity]] | DONE | 2026-06-26 | 2026-06-26 | Convert currency_table.org from the SQL-only ores.codegen.table type to the unified ores.codegen.entity type so that all currency code (C++, SQL, Qt, NATS) can be generated from a single model. |
 | [[id:78BF95FE-1513-4556-BCBC-8A4A1AF48907][Relocate random-pick generator logic out of currency Qt UI]] | DONE | 2026-07-05 | 2026-07-06 | CurrencyDetailDialog::onGenerateClicked() and CurrencyMdiWindow::generateSynthetic() do inline std::uniform_int_distribution random-pick selection over generated currencies; this belongs in the generator/service layer, not the Qt UI. Relocate it before the Qt sync task touches these files, so regeneration wires to the relocated call rather than re-embedding the logic. |
-| [[id:2EAF2F99-B71F-410D-985A-7A616E40CB88][Land CSV/XML import-export as a qt-profile template capability]] | BACKLOG | | | Currency's CurrencyMdiWindow has hand-rolled CSV export and XML import/export with no template equivalent. Generalise it into an opt-in qt-profile mdi_window template capability, with currency as the first real consumer, so the Qt sync task doesn't silently drop it on regeneration. |
+| [[id:2EAF2F99-B71F-410D-985A-7A616E40CB88][Land CSV/XML import-export as a qt-profile template capability]] | STARTED | 2026-07-06 | | Currency's CurrencyMdiWindow has hand-rolled CSV export and XML import/export with no template equivalent. Generalise it into an opt-in qt-profile mdi_window template capability, with currency as the first real consumer, so the Qt sync task doesn't silently drop it on regeneration. |
 | [[id:D9CB6853-797A-4144-BED0-AEBDDBBC173C][Land setting-gated action-visibility knob as a qt-profile template capability]] | DONE | 2026-07-06 | 2026-07-06 | Currency's feature-flagged 'generate synthetic test data' button (gated by system.synthetic_data_generation) has no template equivalent. Generalise it as a conditional action-visibility knob in the mdi_window and detail_dialog templates, with currency as the first real consumer. |
 | [[id:DBDECEFC-7154-4A40-A554-C897AA51BEEE][Land version-navigation UI as a qt-profile detail_dialog template capability]] | BACKLOG | | | Currency's CurrencyDetailDialog has a full version-navigation UI (first/prev/next/last/revert) 100% absent from the template and the synced peer. Generalise it as a detail_dialog template capability, with currency as the first real consumer. |
 | [[id:D613BDFD-FE12-4945-8878-5C92FA109E40][Land NATS notification-wiring capability in the qt-profile controller template]] | BACKLOG | | | Currency's CurrencyController has hand-added onNotificationReceived staleness-propagation logic; the template's changed_event_class flag wires the subscription but generates no handler. Land the handler-generation half of this capability, with currency as the first real consumer (no peer currently configures changed_event_class). |
