--- projects/ores.qt/refdata/include/ores.qt/ImportCurrencyDialog.hpp
@@ -1,161 +1,0 @@
-/* -*- mode: c++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*-
- *
- * Copyright (C) 2025 Marco Craveiro <marco.craveiro@gmail.com>
- *
- * This program is free software; you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation; either version 3 of the License, or (at your option) any later
- * version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
- * details.
- *
- * You should have received a copy of the GNU General Public License along with
- * this program; if not, write to the Free Software Foundation, Inc., 51
- * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
- *
- */
-#ifndef ORES_QT_IMPORT_CURRENCY_DIALOG_HPP
-#define ORES_QT_IMPORT_CURRENCY_DIALOG_HPP
-
-#include "ores.logging/make_logger.hpp"
-#include "ores.qt/ClientManager.hpp"
-#include "ores.refdata.api/domain/currency.hpp"
-#include <QCheckBox>
-#include <QDialog>
-#include <QFutureWatcher>
-#include <QLabel>
-#include <QProgressBar>
-#include <QPushButton>
-#include <QString>
-#include <QTableWidget>
-#include <atomic>
-#include <vector>
-
-namespace ores::qt {
-
-/**
- * @brief Dialog for previewing and importing currencies from XML files.
- *
- * Provides a table view of all currencies to be imported with:
- * - Checkboxes for selecting which currencies to import
- * - Full preview of currency details
- * - Progress bar showing import progress
- * - Status updates during import
- * - Cancel capability during import
- */
-class ImportCurrencyDialog final : public QDialog {
-    Q_OBJECT
-
-private:
-    inline static std::string_view logger_name = "ores.qt.import_currency_dialog";
-
-    [[nodiscard]] static auto& lg() {
-        using namespace ores::logging;
-        static auto instance = make_logger(logger_name);
-        return instance;
-    }
-
-public:
-    /**
-     * @brief Constructs the import dialog.
-     * @param currencies Currencies parsed from XML file
-     * @param filename Path to the XML file being imported
-     * @param clientManager Client manager for sending currencies to server
-     * @param parent Parent widget
-     */
-    explicit ImportCurrencyDialog(const std::vector<refdata::domain::currency>& currencies,
-                                  const QString& filename,
-                                  ClientManager* clientManager,
-                                  const QString& username,
-                                  QWidget* parent = nullptr);
-
-    ~ImportCurrencyDialog() override;
-
-    /**
-     * @brief Get the currencies selected for import.
-     * @return Vector of selected currencies
-     */
-    [[nodiscard]] std::vector<refdata::domain::currency> getSelectedCurrencies() const;
-
-signals:
-    /**
-     * @brief Emitted when import completes successfully.
-     * @param success_count Number of currencies successfully imported
-     * @param total_count Total number of currencies attempted
-     */
-    void importCompleted(int success_count, int total_count);
-
-    /**
-     * @brief Emitted when import is cancelled.
-     */
-    void importCancelled();
-
-private slots:
-    /**
-     * @brief Handles the Import button click.
-     */
-    void onImportClicked();
-
-    /**
-     * @brief Handles the Cancel button click.
-     */
-    void onCancelClicked();
-
-    /**
-     * @brief Handles Select All checkbox state change.
-     */
-    void onSelectAllChanged(int state);
-
-    /**
-     * @brief Handles individual currency checkbox state change.
-     */
-    void onCurrencyCheckChanged();
-
-private:
-    /**
-     * @brief Sets up the UI components.
-     */
-    void setupUI();
-
-    /**
-     * @brief Populates the table with currency data.
-     */
-    void populateTable();
-
-    /**
-     * @brief Updates the Import button enabled state.
-     */
-    void updateImportButtonState();
-
-    /**
-     * @brief Updates the selection count label.
-     */
-    void updateSelectionCount();
-
-private:
-    std::vector<refdata::domain::currency> currencies_;
-    std::vector<std::string> validation_errors_;
-    QString filename_;
-    ClientManager* clientManager_;
-    QString username_;
-
-    // UI Components
-    QLabel* filenameLabel_;
-    QCheckBox* selectAllCheckbox_;
-    QLabel* selectionCountLabel_;
-    QTableWidget* currencyTable_;
-    QProgressBar* progressBar_;
-    QLabel* statusLabel_;
-    QPushButton* importButton_;
-    QPushButton* cancelButton_;
-
-    bool importInProgress_;
-    std::atomic<bool> cancelRequested_;
-};
-
-}
-
-#endif
